Consolidate parameterized tests
